chore: refactor MbedTlsSession
