+++ services/network/ConnectionMbedTls.cpp
@@ -15,278 +15,6 @@
 
 namespace services
 {
-    MbedTlsSession::MbedTlsSession(uint32_t identifier)
-        : session({})
-        , identifier(identifier)
-    {
-        mbedtls_ssl_session_init(&session);
-    }
-
-    MbedTlsSession::MbedTlsSession(uint32_t identifier, infra::BoundedConstString hostname)
-        : hostname(hostname)
-        , session({})
-        , identifier(identifier)
-    {
-        mbedtls_ssl_session_init(&session);
-    }
-
-    MbedTlsSession::MbedTlsSession(uint32_t identifier, IPAddress address)
-        : address(address)
-        , session({})
-        , identifier(identifier)
-    {
-        mbedtls_ssl_session_init(&session);
-    }
-
-    MbedTlsSession::MbedTlsSession(network::MbedTlsPersistedSession& reference)
-        : address(services::ConvertFromUint32(reference.address))
-        , hostname(reference.hostname)
-        , clientSessionObtained(reference.clientSessionObtained)
-        , session({})
-        , identifier(reference.identifier)
-    {
-        mbedtls_ssl_session_init(&session);
-        mbedtls_ssl_session_load(&session, reference.serializedSession.begin(), reference.serializedSession.size());
-    }
-
-    MbedTlsSession::~MbedTlsSession()
-    {
-        if (clientSessionObtained)
-            mbedtls_ssl_session_free(&session);
-    }
-
-    void MbedTlsSession::Reinitialize()
-    {
-        mbedtls_ssl_session_init(&session);
-    }
-
-    void MbedTlsSession::Obtained()
-    {
-        clientSessionObtained = true;
-    }
-
-    bool MbedTlsSession::IsObtained()
-    {
-        return clientSessionObtained;
-    }
-
-    int MbedTlsSession::SetSession(mbedtls_ssl_context* context)
-    {
-        return mbedtls_ssl_set_session(context, &session);
-    }
-
-    int MbedTlsSession::GetSession(mbedtls_ssl_context* context)
-    {
-        return mbedtls_ssl_get_session(context, &session);
-    }
-
-    IPAddress& MbedTlsSession::Address()
-    {
-        return address;
-    }
-
-    infra::BoundedConstString MbedTlsSession::Hostname()
-    {
-        return hostname;
-    }
-
-    MbedTlsSessionWithCallback::MbedTlsSessionWithCallback(uint32_t identifier, const infra::Function<void(MbedTlsSession*)>& onObtained)
-        : MbedTlsSession(identifier)
-        , onObtained(onObtained)
-    {}
-
-    MbedTlsSessionWithCallback::MbedTlsSessionWithCallback(uint32_t identifier, infra::BoundedConstString hostname, const infra::Function<void(MbedTlsSession*)>& onObtained)
-        : MbedTlsSession(identifier, hostname)
-        , onObtained(onObtained)
-    {}
-
-    MbedTlsSessionWithCallback::MbedTlsSessionWithCallback(uint32_t identifier, IPAddress address, const infra::Function<void(MbedTlsSession*)>& onObtained)
-        : MbedTlsSession(identifier, address)
-        , onObtained(onObtained)
-    {}
-
-    MbedTlsSessionWithCallback::MbedTlsSessionWithCallback(network::MbedTlsPersistedSession& reference, const infra::Function<void(MbedTlsSession*)>& onObtained)
-        : MbedTlsSession(reference)
-        , onObtained(onObtained)
-    {}
-
-    void MbedTlsSessionWithCallback::Obtained()
-    {
-        MbedTlsSession::Obtained();
-        onObtained(this);
-    }
-
-    MbedTlsSession* MbedTlsSessionStorageRam::NewSession(infra::BoundedConstString hostname)
-    {
-        storage.emplace_back(0, hostname);
-        return &storage.back();
-    }
-
-    MbedTlsSession* MbedTlsSessionStorageRam::NewSession(IPAddress address)
-    {
-        storage.emplace_back(0, address);
-        return &storage.back();
-    }
-
-    void MbedTlsSessionStorageRam::Invalidate(MbedTlsSession* sessionToInvalidate)
-    {
-        for (auto& session : storage)
-        {
-            if (sessionToInvalidate == &session)
-                storage.remove(session);
-        }
-    }
-
-    MbedTlsSession* MbedTlsSessionStorageRam::GetSession(infra::BoundedConstString hostname)
-    {
-        for (auto& session : storage)
-        {
-            if (session.Hostname() == hostname)
-                return &session;
-        }
-        return nullptr;
-    }
-
-    MbedTlsSession* MbedTlsSessionStorageRam::GetSession(IPAddress address)
-    {
-        for (auto& session : storage)
-        {
-            if (session.Address() == address)
-                return &session;
-        }
-        return nullptr;
-    }
-
-    bool MbedTlsSessionStorageRam::Full() const
-    {
-        return storage.full();
-    }
-
-    void MbedTlsSessionStorageRam::Clear()
-    {
-        storage.clear();
-    }
-
-    MbedTlsSessionStoragePersistent::MbedTlsSessionStoragePersistent(services::ConfigurationStoreAccess<infra::BoundedVector<network::MbedTlsPersistedSession>> nvm)
-        : nvm(nvm)
-    {
-        LoadSessions();
-    }
-
-    MbedTlsSession* MbedTlsSessionStoragePersistent::NewSession(infra::BoundedConstString hostname)
-    {
-        storage.emplace_back(NextIdentifier(), hostname, [this](MbedTlsSession* session)
-            {
-                SessionUpdated(session);
-            });
-        return &storage.back();
-    }
-
-    MbedTlsSession* MbedTlsSessionStoragePersistent::NewSession(IPAddress address)
-    {
-        storage.emplace_back(NextIdentifier(), address, [this](MbedTlsSession* session)
-            {
-                SessionUpdated(session);
-            });
-        return &storage.back();
-    }
-
-    void MbedTlsSessionStoragePersistent::Invalidate(MbedTlsSession* sessionToInvalidate)
-    {
-        for (auto& session : storage)
-        {
-            if (sessionToInvalidate == &session)
-                storage.remove(session);
-        }
-        // nvm.Write();
-    }
-
-    MbedTlsSession* MbedTlsSessionStoragePersistent::GetSession(infra::BoundedConstString hostname)
-    {
-        for (auto& session : storage)
-        {
-            if (session.Hostname() == hostname)
-                return &session;
-        }
-        return nullptr;
-    }
-
-    MbedTlsSession* MbedTlsSessionStoragePersistent::GetSession(IPAddress address)
-    {
-        for (auto& session : storage)
-        {
-            if (session.Address() == address)
-                return &session;
-        }
-        return nullptr;
-    }
-
-    bool MbedTlsSessionStoragePersistent::Full() const
-    {
-        return nvm->full();
-    }
-
-    void MbedTlsSessionStoragePersistent::Clear()
-    {
-        storage.clear();
-        nvm->clear();
-        nvm.Write();
-    }
-
-    uint32_t MbedTlsSessionStoragePersistent::NextIdentifier()
-    {
-        return nvm->size();
-    }
-
-    void MbedTlsSessionStoragePersistent::SessionUpdated(MbedTlsSession* session)
-    {
-        size_t outputLen;
-        network::MbedTlsPersistedSession* persistedSession = nullptr;
-        for (auto& nvmSession : *nvm)
-        {
-            if (nvmSession.identifier == session->identifier)
-                persistedSession = &nvmSession;
-        }
-        if (persistedSession == nullptr)
-        {
-            if (nvm->full())
-                nvm->erase(nvm->begin());
-            nvm->emplace_back();
-            persistedSession = &nvm->back();
-        }
-        persistedSession->address = services::ConvertToUint32(session->address.Get<IPv4Address>());
-        persistedSession->hostname = session->hostname;
-        persistedSession->clientSessionObtained = session->clientSessionObtained;
-        persistedSession->identifier = session->identifier;
-        persistedSession->serializedSession.resize(persistedSession->serializedSession.max_size());
-        auto result = mbedtls_ssl_session_save(&session->session, persistedSession->serializedSession.begin(), persistedSession->serializedSession.max_size(), &outputLen);
-        persistedSession->serializedSession.resize(outputLen);
-        // really_assert(result == 0);
-        if (result != 0)
-            services::GlobalTracer().Trace() << "<<<<<< SSL Session Save Resulted in Failure >>>>>>" << outputLen;
-        // really_assert(outputLen < persistedSession->serializedSession.max_size());
-        if (outputLen > persistedSession->serializedSession.max_size())
-            services::GlobalTracer().Trace() << "<<<<<< SSL Session Save Does not fit into buffer >>>>>>" << outputLen;
-        nvm.Write();
-
-        // remove
-        services::GlobalTracer().Trace() << "Updating Persistent Session with size: " << outputLen;
-    }
-
-    void MbedTlsSessionStoragePersistent::LoadSessions()
-    {
-        for (auto& sector : *nvm)
-        {
-            if (!sector.serializedSession.empty())
-            {
-                storage.emplace_back(sector, [this](MbedTlsSession* session)
-                    {
-                        SessionUpdated(session);
-                    });
-            }
-        }
-    }
-
     ConnectionMbedTls::ConnectionMbedTls(infra::AutoResetFunction<void(infra::SharedPtr<services::ConnectionObserver> connectionObserver)>&& createdObserver, CertificatesMbedTls& certificates,
         hal::SynchronousRandomDataGenerator& randomDataGenerator, const ParametersWorkaround& parameters)
         : createdObserver(std::move(createdObserver))
@@ -313,7 +41,6 @@ namespace services
         mbedtls_ssl_conf_rng(&sslConfig, mbedtls_ctr_drbg_random, &ctr_drbg);
         mbedtls_ssl_conf_authmode(&sslConfig, GetAuthMode(parameters));
         mbedtls_ssl_conf_session_tickets(&sslConfig, MBEDTLS_SSL_SESSION_TICKETS_ENABLED);
-        // mbedtls_ssl_conf_renegotiation(&sslConfig, MBEDTLS_SSL_RENEGOTIATION_ENABLED);
 
         certificates.Config(sslConfig);
 
@@ -916,7 +643,7 @@ namespace services
         if (savedSession == nullptr)
         {
             if (sessionStorage.Full())
-                sessionStorage.Invalidate(session);
+                session != nullptr ? sessionStorage.Invalidate(session) : sessionStorage.Clear();
             savedSession = sessionStorage.NewSession(address);
         }
         session = savedSession;
@@ -989,10 +716,12 @@ namespace services
     infra::SharedPtr<ConnectionMbedTls> ConnectionFactoryWithNameResolverMbedTls::Allocate(infra::AutoResetFunction<void(infra::SharedPtr<services::ConnectionObserver> connectionObserver)>&& createdObserver, infra::BoundedConstString hostname)
     {
         auto savedSession = sessionStorage.GetSession(hostname);
+        // remove
+        services::GlobalTracer().Trace() << "Session: " << hostname << " exists: " << (savedSession == nullptr ? "0" : "1");
         if (savedSession == nullptr)
         {
             if (sessionStorage.Full())
-                sessionStorage.Invalidate(session);
+                session != nullptr ? sessionStorage.Invalidate(session) : sessionStorage.Clear();
             savedSession = sessionStorage.NewSession(hostname);
         }
